// src/components/domains/danji/detail/danji-around-map-card/types.ts
export type AroundListType = {
  address_name: string;
  category_group_code: string;
  category_group_name: string;
  category_name: string;
  distance: string;
  place_name: string | string[];
  x: string | string[];
  y: string | string[];
  id: string;
  phone: string;
  road_address_name: string;
  place_url: string;
}[];

export type Item = {
  id: string;
  place_name: string | string[];
  address_name: string;
  category_group_code: string;
  category_group_name: string;
  category_name: string;
  distance: string;
  x: string | string[];
  y: string | string[];
  phone: string;
  road_address_name: string;
  place_url: string;
};